requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
execnet==2.0.2
Faker==18.13.0
flake8==6.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
PySocks==1.7.1
pytest==7.4.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==3.8.6
wsproto==1.2.0
